reverse changes to configurator & mconfig_digest
